==> Testing Phase/02 - Expected Outcome Updated.ipynb <==
from openai import OpenAI
import os
import pandas as pd
import json

client = OpenAI(
    api_key=os.environ.get("OPENAI_API_KEY"))

#API confing
with open("Evaluation_Prompt.json", "r") as file:
    data = json.load(file)
    
texts = data["texts"]
prompts = data["prompts"]

#Models to tests
models = ["gpt-4o", "gpt-4o-mini", "o1-mini", "gpt-4-turbo"]

#Initialazing list to store results
results = []

#Combinations loop
for model in models:
    for prompt in prompts:
        for text in texts:
            for repetition in range(5):
                try:
                    #Model input
                    message = prompt + text

                    #OpenAI API calling
                    if model == "o1-mini":
                        response = client.chat.completions.create(
                            model=model,
                            messages=[{"role": "user", "content": message}]
                            )
                    else:
                        response = client.chat.completions.create(
                            model=model,
                            messages=[{"role": "user", "content": message}],
                            temperature=0.5
                            )

                    # storing results
                    results.append({
                        "model": model,
                        "prompt": prompt,
                        "text": text,
                        "repetition": repetition + 1,
                        "output": response.choices[0].message.content
                    })

                except Exception as e:
                    #Ignoring errors
                    results.append({
                        "model": model,
                        "prompt": prompt,
                        "text": text,
                        "repetition" : repetition + 1,
                        "error": str(e)
                    })

print("Iteration completed successfully")

# Saving results in JSON file
with open("Iteration_results.json", "w") as file:
    json.dump(results, file, indent=4)

#Dataframing results
df = pd.DataFrame(results)
df.to_csv("Iteration_results.csv", index=False)
df.to_excel("Iteration_results.xlsx")

print("Iterarion result saved. Results saving in 'Iteration_results.json', 'Iteration_results.csv', and 'Iteration_results.xlsx'")

df

# Extract required columns
Results = df[['model', 'text', 'output']]

# Extracting Text Id
Results['text'] = df['text'].str[:7]

# Extracting Suggestion

Results['output'] = df['output'].str.extract(r'\*\*Suggestion\*\*:\s*(Highly appropriate|Mostly appropriate|Partially appropriate|Inappropriate|Highly inappropriate)')[0]
Results

Results.to_csv('Results.csv')
Results.to_excel('Results.xlsx')

Results = pd.read_csv('Results.csv', index_col=0)

Results

expected_output = Results.groupby('text')['output'].describe().reset_index()[['text','top']]

#Deleting blank caracters
expected_output['text'] = expected_output['text'].str.strip()
expected_output['top'] = expected_output['top'].str.strip()
expected_output

expected_output.to_csv('Expected_Outout.csv')

expected_output.info()
