==> Testing Phase/03 - Testing - Updated.ipynb <==
#Importing libraries
import pandas as pd
import json

#Defining texts and prompts
texts = [
    """Text 1 : Dairy beef could help cut greenhouse gas emissions by up to 48% - AgResearch. A new study by AgResearch found that beef from New Zealand’s dairy sector cattle could provide up to 48% reductions in greenhouse gas emissions, compared to the average for beef cattle. The Life Cycle Assessment analysis by AgResearch scientists is supported by the Bioresource Processing Alliance and in partnership with organisations such as Fonterra and Alps 2 Ocean Foods. It shows using fast-finishing systems with dairy-beef animals results in a carbon footprint 32 to 48% lower than the average for traditional beef systems in New Zealand. Farm data from the beef and dairy industries was used in the analysis, which was based on kilograms of carbon dioxide equivalent emissions per kilogram of finished liveweight of the animals. Scenarios were analysed for dairy-beef steers processed between 10 and 18 months of age and included emissions from animal production, input generation, transportation and background processes. AgResearch senior scientist Andre Mazzetto said there were “many issues at play” in determining those emissions and the differences between the dairy and beef animals. “These include how emissions are allocated between milk and liveweight, and issues such as growth rates, feed efficiency and the finishing period/time to slaughter of the animals.” Mazzetto said the work pointed to tangible ways farmers could reduce their beef herds’ carbon footprint. “Emissions reductions can be achieved simply by integrating dairy-beef animals into the herd, but the scope of this reduction can be significantly increased with fast-finishing beef systems. “There would be few tools or management strategies available which can have such significant climate impact.” Mazzetto said sourcing calves from the dairy industry presented an opportunity for the beef industry to reduce its emissions and provided an alternative use for calves from the dairy industry. “There are efforts under way to find new avenues for use of these young dairy-beef animals, such as Alps 2 Ocean Food’s Mīti product made from the meat. “However, we recognise there are also management and logistical challenges for the industries to overcome, including the finishing and processing of these animals.”""",
    """Text 2 : Sex worker Whitney Iraia-Burgess jailed for killing Daniel Otess in Grafton, Auckland, as judge rejects rape claim. An Auckland Central sex worker has been sentenced to three years and three months’ jail for manslaughter after she lashed out at a father-of-two with a knife in an attack she claims followed being raped in the back of a car. However, at a disputed facts hearing prior to sentencing in Auckland High Court today, Crown prosecutor Brett Tantrum argued there was no rape in an open-air Grafton car park in the early hours of March 12, 2023, and the attack was fuelled by Whitney Iraia-Burgess, 33, seeking money for her meth habit. Justice Geoffrey Venning agreed with the Crown there was no rape or attempted rape and “at most” there was an argument with the victim Daniel Otess, 57, about money in the City Rd carpark that “initially led to pushing and shoving” outside the car. Iraia-Burgess then slashed at Otess with a knife, which her defence barrister Mark Ryan said was carried for self-defence as a working girl, resulting in five sharp force injuries – one of which severed an artery in Otess’ arm resulting in “massive” blood loss. He died beside his car in a pool of blood as Iraia-Burgess fled with $100, with two unused condoms left at the scene. Then-murder suspect Iraia-Burgess was the subject of a multi-day manhunt last year after Otess’ body was found. The father of two was a night cleaner originally from Ghana whose death was described by those who knew him as having “crushed the hearts of families in two countries”. North Shore resident Iraia-Burgess stood in the dock in the High Court at Auckland this morning as she admitted the fatal stabbing of Otess – but claimed she had begun stabbing him in the back seat in an attempt to escape as he had pinned her arm while raping her. “There was a fight in the back seat where he had his hand around my wrist … that’s why I cut his hand because I couldn’t get out of the vehicle at the time,” she said. “What I was trying to do was try to escape. I wasn’t sitting watching to put his pants on … he came out of the vehicle pretty quickly. But Crown prosecutor Tantrum showed CCTV footage from the carpark on the night which he said did not support Iraia-Burgess’ narrative. Tantrum argued there would have been no time for Otess to buckle up his pants and chase after Iraia-Burgess, and there was very little blood found inside his car. Iraia-Burges had been set to go to trial for murder in the middle of this year, but that charge was withdrawn in lieu of Iraia-Burgess’ manslaughter admission. In sentencing today at Auckland High Court, Justice Venning said Iraia-Burgess’ response to the carpark altercation with a knife was “not reasonable”. “The taking of a life in such a senseless way should be condemned,” Venning said. “The long-lasting effect of your offending on Mr Otess’ family should be very evident to you by now.” Victim impact statements read by both of Otess’ daughters in court today described a home “that has become a place of sorrow and emptiness” since the death. “He weighed no more than 55kg. He travelled 10,000 miles leaving behind friends and the only world he knew. He worked hard every day and always gave back to his family at home,” one of Otess’ daughters said. “I am heartbroken. When I wake in the mornings he is no longer there … when I call he no longer answers. There will be no more long drives with my passenger prince. “I will go through a lifetime of missing my father for the privilege of knowing him.” Among a large contingent of Otess’ family in court today was his brother from Ghana. “He considered New Zealand his home,” he said. Relatives in Ghana finding out how Otess died was a “shattering” experience that had deprived them of financial sponsorship from Otess’ devoted work in New Zealand, he said. Two other witnesses who interacted with Iraia-Burgess in the hours prior and following the killing appeared as witnesses in a disputed facts hearing this morning. Tamehana Maru, who appeared via video link, was with Iraia-Burgess prior to meeting Otess on the night of March 11, 2023. The pair were smoking meth and drinking together. Iraia-Burgess claimed Maru provided her with the knife Otess was killed with when he dropped her off in the middle of the night outside a Countdown. Maru disputed he provided the knife. Iraia-Burgess’ defence barrister Mark Ryan explained the carrying of the knife in court today: “People in this profession are extremely vulnerable and it’s not usual for them to carry something for their protection”. Iraia-Burgess admitted she had carried knives before but not used them. She had also had two prior offences for knife possession in 2016 and 2019. Another witness appearing by video link today in court was Joshua Wong who met up with Iraia-Burgess after the stabbing of Otess. Wong denied helping Iraia-Burgess dispose of the knife which was wrapped up and thrown off a bridge, and taking her clothes and one of the cellphones she had on her that night. “He did, he guided me on what to do the whole time. I was in shock … I didn’t know what to do because I was being tracked. I had my phones on me,” Iraia-Burgess claimed of Wong. Wong also disputed saying “The police will be here any minute to raid your house if he dies” to Iraia-Burgess on the night of the attack. “She said when she got in the car that she’d been in a fight. That was going to be the best response to use for why she did what she did,” Wong said of what Iraia-Burgess had told him that night. In sentencing today, Justice Venning gave discounts from his starting point of six years’ jail for Iraia-Burgess’ extremely difficult upbringing, efforts at rehibilitation, guilty plea for manslaughter and time already served on bail. “You’re now 33 years old. Your life has been a tragic one. You were abandoned by your mother shortly after your birth,” Venning said, adding she was “a victim of physical and sexual abuse” during her youth. Iraia-Burgess’ was supported in court by her teenage daughter today.Carpark confrontation Police launched an investigation last year after a group found Otess’ body beside a car in the open-air carpark on Grafton’s City Rd, near Symonds St around 5.20am on March 12, 2023. Police quickly identified Iraia-Burgess as the suspected culprit and sought the public’s help in locating her. She handed herself in after several days in hiding, authorities said at the time. “Ms Iraia-Burgess and Mr Otess were not known to each other prior to the incident,” authorities noted in the summary of facts for the case, a document that was agreed to by both the Crown and the defence. “At about 4.20am ... [Otess’] vehicle stopped nearby where Ms Iraia-Burgess was standing on Liverpool Street. “The two had a quick conversation via the vehicle window and Ms Iraia-Burgess then got into the front passenger seat of Mr Otess’ vehicle, seemingly for her services of sexual activity.” Police noted soon after the homicide investigation began that Otess had been captured on CCTV at a nearby petrol station an hour before the discovery of his body. It was revealed today that the defendant had been with him. He withdrew $100 before returning to the car. After arriving at the “secluded” carpark, the pair parked diagonally in a corner. “Soon after their arrival at 4.44am, Ms Iraia-Burgess and Mr Otess started a physical altercation outside Mr Otess’ vehicle,” court documents state. “Ms Iraia-Burgess pulled out an unknown weapon that was already in her possession. She slashed Mr Otess multiple times around the head, shoulder and the arm area.” Suffering five “sharp-force injuries”, as well as blunt-force injuries to his nose and surrounding areas, Otess attempted to get back into his vehicle after the defendant fled the scene on foot with the cash he had withdrawn. “However, due to massive blood loss from the wound to his left forearm, he collapsed to the ground and died next to his vehicle,” the documents state. “The cause of death was the sharp force injury to the left arm.” Family-oriented victim, religious defendant The violent last moments of Otess’ life stand in stark contrast to the man remembered last year by friends in Auckland’s tight-knit Ghanaian community, which was left reeling after as it came to grips with the limited information about what had happened. “New Zealand is considered by our community as a safe and peaceful country to live in but this incident has shaken that confidence,” Charles Ampomah-Dwamena, who serves as the West African nation’s honorary consul in New Zealand, said at the time. Friends remembered Otess as a family man and a “very calm person” who wasn’t known to seek confrontation. “He will be missed by the whole community, he was a part of us,” friend Bright Duah, a member of the Ghana Association of New Zealand, previously told the Herald. “He was a lovely and caring family man. He had recently taken his family to Ghana to meet his parents there.”""",
    """Text 3 : Taylor Swift started the Eras Tour when people needed it most. Now, it’s time to say goodbye. Taylor Swift was still a new-to-the-scene country artist when she performed at Cowboys, a small bar in the Colorado Springs area, in 2007. At the end of the show, she took two purple beaded bracelets off of her wrist and handed them to a woman who’d been at the front of the stage, singing her songs louder than anyone else. Then-college student Carola Lovering, the recipient of Swift’s token of appreciation, kept one bracelet for herself and gave the other to her best friend, Joanie Goldfarb, who was standing next to her. Lovering, who is now an established author and mom of two, and Goldfarb reunited at the Eras Tour this October. Though the bracelets once gifted to them by Swift were long lost, they exchanged new ones with other concertgoers, taking part in a now-tradition born during the superstar’s Eras Tour, the final three shows of which begin Friday in Vancouver. Everyone from high-ranking politicians to Sir Paul McCartney himself showed up to see Swift perform a three-hour show that included a 44-song setlist of her greatest hits spanning her now 18-year career. Her concerts caused earthquakes, created a Beatlemania-level of fervor, impacted local economies and brought a collective joy to millions of people still reeling from the dark days of the pandemic. As the tour wraps up after a two-year run around the globe, this is not just the end of an important chapter in Swift’s career, it’s a momentous conclusion to a significant chapter in modern music history. “I think at an emotional and cultural level, it’s on par with, let’s say Elvis, Michael Jackson, Sinatra, you could say Madonna, and also the Beatles. Just that deep attachment to these artists and the frenzy around it,” Jasen Emmons, chief curator and VP of curatorial affairs at Los Angeles’ Grammy Museum told CNN in a recent interview. “Whether she orchestrated it or not, she’s created this zeitgeist around that tour that feels unprecedented.” Embracing togetherness As were so many other industries, the music world was rocked after the global pandemic shut down restaurants, theaters, music venues or anywhere people would gather for a good time. Swift was set to kick off her previously announced Lover Fest tour when that was postponed and then ultimately cancelled. One year later, she announced the Eras Tour, her first tour in six years. The Eras Tour began in March 2023, just as life was coming back into full view. “Coming out of the pandemic there was such a pent up demand to go out and see something, especially something that felt ‘this is going to be so special that I want to be a part of it’ because you realize ‘I’m part of a larger cultural event,’” said Emmons. You could say the Eras Tour came at a time when people needed it most. The same has been said about the Beatles, who made their first visit to the US in February 1964, just months after the assassination of President John F. Kennedy. In “Beatles ’64,” Paul McCartney acknowledged that the timing of the tragedy and Beatlemania in the US were linked. “Maybe America needed something like The Beatles to be lifted out of sorrow,” he said. Swift similarly acknowledged this sentiment in her commemorative Eras Tour book when she wrote in the forward that she created the tour, in part, “because people need an escape from how brutal life can be.” That desire to be part of the healing fervor of The Eras Tour became apparent early on when hundreds of ticketless Swifties would gather in parking lots or vacant fields near several concert venues on show night for what became known as “Taylor-gating,” where they would trade bracelets and sing along to the live show they could hear but not see. “Despite the vastness of what she’s doing each night, say, playing to 80,000 people, she has this ability to make it feel intimate,” Emmons said. “So people, when they go, feel this direct connection to her.” The cost of a memorySwift’s team has not released official figures on how much money the Eras Tour grossed but it’s been estimated that the tour made over $2 billion, making it the highest-grossing tour in history. The current record-holder is Elton John, whose Farewell Yellow Brick Road tour grossed over $900 million. It’s a feat considering when tickets first went on sale, the demand was so high that Ticketmaster crashed. According to Tom Corbett, Global Head of Sponsorship at Barclays in the UK, the Eras Tour has highlighted the rise of consumers who are prioritizing spending their money on memorable experiences instead of spending on physical items. “The Eras tours has been the poster child of the ‘experience economy’ that we talk about,” Corbett told CNN in a recent interview. “We’re seeing a shift in people’s behaviors and I think that’s going to continue.” In May, Barclays estimated that the Eras Tour was set to boost spending in the United Kingdom by around $1.2 billion (£1 billion) in June, when her tour would hit cities across England, Wales and Scotland. Based on their data, Corbett said, those estimates lived up to their expectations. “We’re pretty confident in that figure,” he said. “Certainly the Eras Tour, the brand of Taylor Swift, generated over a billion pounds of spending on goods and services from the tickets to beads to eating out to travel, which is kind of unbelievable when you say that number out loud.” The UK isn’t the only economy that the Eras tour affected, with several cities across the US benefiting from local economic boost in what became known as the “Swift Effect.” The price of tickets did however loom large. If you weren’t able to purchase a tickets at face value during the initial presale, chances are you bought tickets at an inflated cost on the resale market. It’s what happened to Lovering, who decided to splurge on a resale ticket that cost just south of $2,000 for the Miami concert she saw in October. Part of the draw has been the show performed throughout the first US leg in 2023 is different from the show she’s performing in 2024, thanks to a mid-tour setlist revamp made to include songs from her 11th studio album “The Tortured Poets Department,” released in April. “We want to be able to tell our kids we went to the Eras Tour,” Lovering told CNN. “Because it’s just never going to happen again like this.”""",
    """Text 4 : Syrian rebels claim to take control of Daraa, birthplace of the 2011 uprising, inching closer to Damascus. Syrian opposition forces claim to have taken control of Daraa city in southwestern Syria, inching closer to the capital Damascus. The rebels are now battling Syrian President Bashar al-Assad’s forces from two directions; the north and the south, in an attempt to close in on Damascus. Daraa is where the Syrian uprising started in 2011. The Syrian Defense Ministry has not yet confirmed or denied that rebels are in control of Daraa. “Our forces have taken full control of the entire city of Daraa and have started combing through its neighborhoods and securing its institutions and government offices,” local rebel fighters known as the Southern Operations Room said in a statement late Friday. A video, geolocated by CNN, shows rebels outside the Daraa administrative building. The southern factions are different from factions in the north led by the Islamist rebel group Hayat Tahrir al-Sham (HTS), who have already taken two major cities during their offensive that started over a week ago. Still, those groups aim for the same goal, which is toppling Assad. The offensive has reawakened a civil war that had been largely dormant for years. The conflict began in 2011, after Assad moved to stamp out peaceful pro-democracy protests during the Arab Spring. The fighting swelled as other regional actors and world powers – from Saudi Arabia and the United States to Iran and Russia – piled in, escalating the civil war into what some observers described as a “proxy war.” More than 300,000 civilians have been killed in more than a decade of war, according to the United Nations, with millions more displaced across the region. On Friday, rebels in southern Syria also took control of a Syria-Jordan border crossing, after launching a fresh offensive. The Nassib border crossing marks the southernmost point of the major M5 highway, which runs from the northern city of Aleppo and through the capital. In a video geolocated by CNN, armed fighters were seen marching down a road by the Nassib border crossing, after wresting it from government control. The rebels in the north, after capturing Aleppo a week ago, have pushed south along the highway and on Thursday took the city of Hama. They have now set their sights on Homs. And as anti-regime offensives gather pace both to the north and south of the capital Damascus, Kurdish-led fighters say they have moved into eastern parts of Syria. Kurdish forces occupy the northeast of the country, having achieved hard-won autonomy during a decade of civil war. They fear that autonomy could now be under threat by the Syrian opposition insurgents. Although the stated target of the rebel offensive is the Assad regime, Kurdish fighters have said they fear the violence could spill over into their territory. Fear of Retaliation Hundreds of people appear to have fled the central Syrian city of Homs overnight into Friday. Videos showed hundreds of vehicles lining the highway out of Homs, as the city braces for violence. After capturing the city of Hama to the north on Thursday, the rebels set their sights on the crossroads city of Homs which, if captured, would split the territory under the control of the regime in two. “From here we direct the last call to the regime forces, this is your chance to defect,” said the Military Operations Media of the armed opposition groups led by HTS, in a statement. Late on Friday, the rebels claimed to have seized the last village on the outskirts of Homs city, saying they are at the city walls. The goal of Syria’s rebel coalition, which has wrested two major cities from government control in just over a week, is ultimately to overthrow the decades-old Assad regime, the militant leader of HTS, Abu Mohammad al-Jolani, told CNN in an exclusive interview on Thursday. “When we talk about objectives, the goal of the revolution remains the overthrow of this regime. It is our right to use all available means to achieve that goal,” said Jolani. Homs has a sizeable population of Alawites, co-religionists of Assad, many of whom fear retribution from rebels who accuse Alawites of dominating the country and supporting Assad’s quashing of the rebellion. After bursting out of their pocket of territory in the northwest region of Idlib, the rebels captured Aleppo within three days and Hama within eight, meeting minimal resistance from regime forces which were caught flat-footed by the lightning offensive. It is not clear whether regime forces will be able to defend the city of Homs, about an hour’s drive south of Hama. CNN was unable to reach multiple contacts in Homs on Friday, amid reports of an internet blackout as the rebels approached the city. The rebels’ progress has been met with celebration from many residents of the newly captured cities who had lived under regime rule for years. Videos geolocated by CNN showed rebel fighters cheering – almost in disbelief at their progress – as they entered Hama. “Guys, my country is being liberated. I swear to God, we are inside Hama city, we are inside Aleppo city,” a fighter cheered as he filmed himself by Alaarbaen Roundabout in Hama. The rebels led by HTS said they freed hundreds of those “wrongfully detained” from Hama’s central prison. The city was the site of one of Syria’s largest massacres in 1982, when President Hafez al-Assad – the current ruler’s father – ordered his military to quash a revolt. In an interview with CNN, Jolani, the leader of HTS, said his group aimed to create a government based on institutions and a “council chosen by the people.” Although the prospect of the rapid collapse of the Assad regime was scarcely conceivable just over a week ago, Jolani said: “The seeds of the regime’s defeat have always been within it… the Iranians attempted to revive the regime, buying it time, and later the Russians also tried to prop it up. But the truth remains: this regime is dead.” Regional Concerns The foreign ministers of Iran, Iraq, and Syria met in Baghdad Friday and said in a joint statement that the advance of opposition forces in Syria poses “a serious danger to the three countries, threatens the security of their peopl es and the region as a whole.” The three countries described the Syrian armed opposition forces as “terrorists,” stressing that there was “a need for collective action to confront it.” Israel has instructed its forces “to maintain a high level of preparedness and to continuously monitor the developments,” in Syria as rebels advance, the country’s defense ministry said in a statement Friday. The Israeli military “is prepared for any scenario and is determined to protect the citizens of Israel and protect Israel’s security interests at all times,” the statement said. The US on Friday urged its citizens to depart Syria “now while commercial options remain available in Damascus.” “U.S. citizens who choose not to depart Syria or are unable to depart should prepare contingency plans for emergency situations and be prepared to shelter in place for extended periods. The Aleppo International Airport is closed,” the State Department said in a statement, citing the “volatile and unpredictable” security situation.""",
    """Text 5 : Nearly 70% of Gaza war dead verified by UN are women and children. The UN's Human Rights Office has condemned the high number of civilians killed in the war in Gaza, saying its analysis shows close to 70% of verified victims over a six-month period were women and children. The agency said the high number was largely due to Israel's use of weapons with wide-area effects in densely populated areas, although some deaths may have been the result of errant projectiles by Palestinian armed groups. The report said it found unprecedented levels of international law violations, raising concerns about war crimes and other possible atrocity crimes. Israel has in the past said it targets Hamas and takes steps to mitigate risk to civilians by using precise munitions. The BBC contacted the Israel Defense Forces (IDF) for comment in response to Friday's report. The UN agency said it verified the details of 8,119 people killed in Gaza from November 2023 to April 2024. Its analysis found around 44% of verified victims were children and 26% women. The ages most represented among the dead were five to nine-year-olds. About 80% of victims were killed in residential buildings or similar housing, the agency added. The report said the data indicates an apparent indifference to the death of civilians and the impact of the means and methods of warfare. Gaza Hamas-run health ministry, whose figures the UN sees as reliable, has reported a death toll of more than 43,300 people over the past 13 months. Many more bodies are believed to remain under the rubble of bombarded buildings. The health ministry said it obtained full demographic data for a majority of those killed and reported that children account for one in three of that number. UN Human Rights Chief Volker Türk said in a statement that this unprecedented level of killing, and injury of civilians is a direct consequence of the failure to comply with fundamental principles of international humanitarian law. He cited the laws of distinction, which requires warring parties to distinguish between combatants and civilians, proportionality, which prohibits attacks where harm to civilians outweighs military advantage, and precautions in attacks. Türk called for a due reckoning with respect to the allegations of serious violations of international law. The IDF has previously told the BBC in response to criticism that it will continue to act, as it always has done, according to international law". The report also said the way the warring parties have conducted the conflict in Gaza has caused horrific human suffering. The UN said Palestinian armed groups have waged war from densely-populated areas and indiscriminately used projectiles, likely contributing to the death toll, while the IDF has destroyed civilian infrastructure and "left many of those alive, injured, displaced and starving, without access to adequate water, food or healthcare". The situation is worst in north Gaza, which aid groups say has been under siege since early October when Israel launched a new ground offensive against Hamas. The UN said no food aid entered the north during the first two weeks of October. This prompted the US to issue an ultimatum to Israel to increase aid by 12 November or risk losing some military support. Jan Egeland, the head of aid organisation Norwegian Refugee Council, told the BBC on Friday that he saw devastation, despair, beyond belief on a recent visit to Gaza. There is hardly a building that is not damaged. And large areas looked like Stalingrad after the Second World War. You cannot fathom how intense this indiscriminate bombing has been on this trapped population," he said. "It's evident that it is first and foremost children and women who are paying a price for this senseless war, he added. Israel launched its current military offensive in Gaza after Hamas attack on 7 October 2023 that killed 1,200 people in Israel and took 251 hostages back to Gaza.""",
    """Text 6 : The photo of two men standing on a mountain of bison skulls is well known as a symbol of hunting during the colonisation of the US. But there's a more sinister story behind it  with a surprising modern message. Two men in black suits and bowler hats pose with a gravity-defying mound of bison skulls. The 19th-Century image is disturbing  thousands upon thousands of skulls piled in neat rows, towering towards the sky. But beneath the macabre first impression, the photo holds a darker secret still. These skulls aren't just the product of overzealous hunting in the US  and those men aren't hunters, either. The skulls, experts say, are the evidence of an organised, carefully calculated campaign to eradicate the bison, deprive Native Americans of a vital resource, and drive the few communities that survived onto small reservations where they could be controlled by the newly arrived white settlers. This image is an example of colonial celebration of destruction, says Tasha Hubbard, a Cree filmmaker who is an associate professor at the Faculty of Native Studies at the University of Alberta in Canada. Hubbard describes the extermination of the bison as a strategic part of colonial expansion. The eradication of the animal was seen as the taming of the West, of domesticating this wild space that was needed in order for expansion of settlement. The colonial mass slaughter of bison dealt a lasting blow to tribes that relied on the animal for sustenance. In the aftermath, nations reliant on bison fared measurably, permanently worse than nations that were never bison-reliant, for example suffering from higher child mortality than those other nations, according to a comparative study. The study concludes that the loss set the bison nations on a fundamentally different trajectory that continues to this day. Native Americans had hunted bison for centuries. For bison nations, it was part of their primarily nomadic culture and the animals provided them with vital sustenance – meat for food, hides for shelter and clothing, and bones for tools. (In common parlance and historical sources, the animals are often referred to as buffalo, as that's what early settlers called them  though the two are in fact different.) Indigenous peoples across North America relied on the animal, Hubbard says. So to remove that keystone species was to weaponise starvation against indigenous peoples: to weaken us in order to control us and remove us from our territories." Despite the bisons' usefulness, estimates put the Native American hunters' take at less than 100,000 a year, hardly making a dent on the early 1800s population of between 30 and 60 million bison. By 1 January 1889, there were just 456 pure-breed bison left in the US – and 256 of them were in captivity, protected in Yellowstone National Park and a handful of other sanctuaries. The reasons for the mass bison slaughter are numerous: they include the building of three railway lines through the most populous bison areas, which brought new demand for the animal's hide and meat; modern rifles that made killing bison relatively easy; a lack of protective measures which could have curbed hunting. But there was also more sinister, targeted reason for the animals' decline than just an increased demand for bison products – more on this later. And even the settlers' seemingly practical need for bison meat and hide was ultimately intertwined with colonisation and conquest, historians say. "A desire for wealth and power in the form of land ownership, chattel slavery, the drive for unending growth and profit, and the commodification of natural resources is the reason for the intense overhunting of bison and the political and physical attacks on indigenous nationhood and humanity over five centuries," says Bethany Hughes, a member of the Choctaw Nation of Oklahoma, and assistant professor at the University of Michigans department of Native American studies. When the Transcontinental Railroad was completed in 1869, it accelerated the decimation of the species. In 1871, a Pennsylvania tannery developed a method of converting bison hides into commercial leather. Swarms of hide hunters decimated central plains herds with a "shocking rapidity", one study noted. The infamous image of bison skulls was taken at the Michigan Carbon Works, a refinery that processed bones. There, the bison bones were processed into charcoal that the sugar industry used to filter and purify sugar – the bones were also used as glue and fertiliser. "This photo records a remarkably successful business that was built on the waste created by American Western expansion and its accompanying racial logics of Native American inferiority," says Hughes. "Colonialism and capitalism travel together," Hughes adds. "To benefit from and encourage the kind of economic success this company had processing bison bones, [which] were the byproduct of the sometimes-violent tactics of American settler colonial expansion, was to benefit from – and participate in colonial projects that stripped Indigenous peoples of land, nationhood, and culture. "This photo is not a bracing reminder of the harms of colonial pasts. It is an indictment on commercial consumption practices that obscure the material and ethical conditions that make luxuries like refined sugar readily available and seemingly benign." Killing bison was also part of military campaigns that used resource deprivation as a tactical move. It has been well documented that Western army officials sent soldiers to kill bison as a way of depleting Native American resources during the colonisation of the US. An analysis by historian Robert Wooster in his book The Military and United States Indian Policy acknowledges that General Phillip Sheridan, an army officer responsible for the "Total War" strategy against Southern Plains tribes,"recognised that eliminating the buffalo might be the best way to force Indians to change their nomadic habits". Sheridan was recorded telling legislators who were trying to pass laws to protect the dwindling herds: "[Hunters] are destroying the Indians" commissary. And it is a well known fact that an army losing its base of supplies is placed at a great disadvantage…for a lasting peace, let them kill, skin and sell until the buffaloes are exterminated." Sheridan wrote in a letter to a fellow general in 1868: "The best way for the government is to now make [the tribes] poor by the destruction of their stock, and then settle them on the lands allotted to them." Another army official – Lieutenant Colonel Dodge – told a hunter: "Kill every buffalo you can! Every buffalo dead is an Indian gone." The Native American tribes knew what was happening. Satanta, chief of the Kiowas tribe, in the Great Plains,recognised that "to destroy the buffalo meant the destruction of the Indian" – as Billy Dixon, a bison hunter and frontiersman from Texas, recalled in his autobiography. "General Phil. Sheridan, to subdue and conquer the Plains tribes for all time, urged and practiced the very thing that Satanta was fearful might happen," Dixon added. Depriving the Native Americans of their bison meant they were forced to move onto the new reservations the Western army had established for them, in order to grow food to survive. The armys tactics worked. The Kiowa Tribe members were later driven onto a reservation in Oklahoma. Within one generation, the average height of Native Americans who had relied heavily on bison and so were most impacted by the slaughter dropped by more than an inch (2.5cm). By the early 20th Century, child mortality was 16% higher, and the income per capita among bison nations has remains 25% lower compared to nations that werent so reliant on bison. However there has been some debate over the years over the kill-off. How could hunters kill 30 to 60 million animals? That was the question posed by a 2018 study which offered a disease epidemic as the answer. Two diseases in the country at the time – anthrax in Nebraska and Texas tick fever in Montana – would have been "sufficiently deadly to wipe out tens of millions of animals", the study notes. Regardless of the cause, the bison populations never fully recovered and the species is still listed as near-threatened. But in recent years efforts have been made to bring bison back to the Great Plains (they are incredibly important to the prairies ecosystem). In the US governments 2023 Inflation Reduction Act, $25m (£19.7m) was pledged to restore bison across the US. Smaller efforts are already underway: 1,000 bison raised on reserves belonging to environmental non-profit The Nature Conservancy have been returned to their ancestral grazing lands. A restoration project in Montana aims to bring 5,000 bison back to the prairies and tribes have returned 250 bison to their land in a partnership with the National Wildlife Federation. The message behind the striking mount of bison skulls image has been lost over time, adds Hughes, who says the image carries a simplistic message that allows viewers to feel sadness about the past, but does not force them to confront "the ways that colonial and capitalist systems continue to negatively shape our environment and our lives". "More than that, this photo points to the ways that consumers of products are the engine that drives the colonial machine. "If you dehumanise another person or objectify a living being as a "natural resource" you have revealed your own lack of humanity and a misunderstanding of what it means to live in relationship with the world around you," Hughes says. "This is an important message to share with the public because this is an ongoing problem, not a historical one.""",
    """Text 7 : Children as young as 10 will face adult jail time in Australian state. The Australian state of Queensland has passed laws which will see children as young as 10 subject to the same penalties as adults if convicted of crimes such as murder, serious assault and break-ins. The government says the harsher sentencing rules are in response to "community outrage over crimes being perpetrated by young offenders" and will act as a deterrent. But many experts have pointed to research showing that tougher penalties do not reduce youth offending, and can in fact exacerbate it. The United Nations has also criticised the reforms, arguing they disregard conventions on the human rights of children and violate international law. The Liberal National Party (LNP) - which won the state election in October - made the rules a hallmark of its campaign, saying they put the "rights of victims" ahead of "the rights of criminals". "These laws are for every Queenslander who has ever felt unsafe and been a victim of youth crime across our state," Premier David Crisafulli said after parliament passed the bill on Thursday. Leading up to the vote, both sides of politics had claimed that Queensland was in the grips of a youth crime wave, and that a more punitive approach was necessary to combat the issue. But data from the Australian bureau of statistics, shows that youth crime has halved in Queensland across the past 14 years, that it hit its lowest rate in recorded history in 2022, and has remained relatively steady since. Figures from the Queensland Police Service and the Australian Institute of Criminology also demonstrate a clear downward trend. Dubbed by the government as "adult crime, adult time", the new laws list 13 offences which will now be subject to harsher prison sentences when committed by youths, including mandatory life detention for murder, with a non-parole period of 20 years. Previously, the maximum penalty for young offenders convicted of murder was 10 years in jail, with life imprisonment only considered if the crime was "particularly heinous". The laws also remove "detention as a last resort" provisions - which favour non-custodial orders, such as fines or community service, for children rather than incarceration - and will make it possible for judges to consider a child's full criminal history when sentencing. The Queensland Police Union has called the changes "a leap forward in the right direction", while Queensland's new Attorney-General Deb Frecklington says it will give courts the ability to "better address patterns of offending" and "hold people accountable for their actions". But in a summary, Frecklington also noted the changes were in direct conflict with international standards, that Indigenous children would be disproportionately impacted and that more youngsters were likely to be held in police cells for extended periods because detention centres are full. Queensland already has more children in detention than any other Australian state or territory. Premier Crisafulli said on Thursday that although there may be "pressure in the short-term" his government had a long-term plan to "deliver a raft of other detention facilities and different options". Australia's commissioner for children, Anne Hollonds, described the changes as an "international embarrassment". She also accused Queensland's government of "ignoring evidence" which suggests "the younger a child comes into contact with the justice system, the more likely it is that they will continue to commit more serious crimes". "The fact that provisions are targeting our most at-risk children makes this retreat from human rights even more shocking," she said in a statement on Wednesday. Other legal experts, who gave evidence to a parliamentary hearing on the bill last week, said the laws could have unintended consequences for victims, with children being less likely to plead guilty given the tougher sentences, resulting in more trials and longer court delays.""",
    """Text 8 : Starmer and Badenoch's lunchtime sandwich split. Some voters complain that politicians are all the same but now a clear difference has emerged between the two main party leaders. In an interview with The Spectator, Conservative Party leader Kemi Badenoch said: "Lunch is for wimps. I have food brought in and I work and eat at the same time. "There's no time... sometimes I will get a steak... I'm not a sandwich person. I don't think sandwiches are a real food, it's what you have for breakfast." She added that she would "not touch bread if it's moist". Asked about her comments, No 10 said Prime Minister Sir Keir Starmer viewed sandwiches as a "great British institution". "I think he was surprised to hear that the leader of the opposition has a steak brought in for lunch. "The prime minister is quite happy with a sandwich lunch." The spokesman added that Starmer "enjoys a tuna sandwich, and occasionally a cheese toastie". Responding to the dig, Badenoch wrote on social media: "The PM has time to respond to my jokes about lunch… but no time for the farmers who produce our food." Nigel Farage, who is well-known for his boozy lunches, has also weighed into the debate. In a video on social media, the Reform UK leader said he thought lunch was "pretty cool" as he sipped from a glass of red wine at a restaurant table. Jim Winship of the British Sandwich Association, unsurprisingly, hit back at Badenoch's comments. He told Matt Chorley on BBC Radio 5 Live: "Rubbish isn't it? We eat 3.5bn commercially-made sandwiches every year, this is part of our heritage. "The first sandwich was recorded as being made in 1762 by the Earl of Sandwich when he called for some cuts of beef to be brought to him between slices of toast so he could play cards. "Sandwiches have been on the menu for a lot of people ever since then." They may be part of the UK's heritage but Badenoch is not alone in her deep dislike of sandwiches. Earlier this month, actor Anna Maxwell Martin told The Sun: "If I'm on a set and sandwiches are brought on - it's hard for me to even say that word - I'd have to say 'get that out'." Sandwiches can be a surprisingly troublesome subject for British politicians. Badenoch's predecessor as Tory leader, Rishi Sunak, was teased during the general election when ITV's This Morning asked what his favourite meal was and he replied "sandwiches". Former Labour leader Ed Miliband was mocked for slightly struggling to eat a bacon sandwich, in a picture that was later seen as helping to scupper his bid to be PM. And Dominic Raab hit the headlines when the then Tory minister's sandwich preferences were revealed by a former member of his staff. She told the Daily Mirror he ordered the same lunch every day from Pret A Manger. "He has the chicken Caesar and bacon baguette, SuperFruit pot and the Vitamin Volcano smoothie, every day. He is so weird. It's the Dom Raab Special." Raab vehemently denied the claim, joking that he was now getting offers of free sandwiches from Subway.""",
    """Text 9 : Indian teen becomes youngest world chess champion. Indian teenager Gukesh Dommaraju has become the youngest-ever world chess champion after beating defending champion China's Ding Liren in a dramatic turn on Thursday. Dommaraju, 18, is four years younger than the former record-holder, Russian grandmaster Garry Kasparov, who was 22 when he won the title in 1985. The Chennai prodigy has long been a superstar in the chess world, having attained the status of chess grandmaster at the age of 12. But he was seen as the outside challenger going into the final round of the FIDE World Chess Championship, held in Singapore this year. Playing on black, Dommaraju won the game after Ding, who had been in a solid position, made a fatal foolish move that gave up his last powerful piece. His blunder delivered victory to the 18-year-old, who until now had been ranked fifth in the world and second in his own country. The 14-game World Championship competition had been closely watched by chess fans around the world this past fortnight. Going into the final game on Thursday, Dommaraju and Ding had eight draws and two wins apiece. Players receive one point for a win and half a point each for a draw. Dommaraju claimed the title on Thursday with a final score of 7.5 to 6.5, becoming just the 18th world chess champion. The teenager comes from Chennai, a city known as India's chess capital for having produced so many national champions. But there were no elite chess players in his family – he was enrolled in chess sessions after school because his father, a surgeon, and his mother, a medical professor, needed somewhere to put him. His talent was spotted there by coaches, who encouraged his family to invest in his training. In high school in 2019, he was crowned a grandmaster at the age of 12 years and seven months – the third-youngest in history. The teenager has spoken before about how yoga and mindful thinking has helped him deal with the pressures of his chess career. He stayed focused on Thursday as his opponent, the defending champion Ding, appeared to buckle under the pressure. Ding has faced questions over his form all year since winning the title in 2023 becoming China's first chess world champion. For most of the year, he had taken a break from chess, having spoken about his struggles with depression and mental health. But his stylish win over Dommaraju in the opening game of the championship last month, and a victory in Round 12, had suggested momentum. Thursday's game saw several hours of tight play, with commentators suggesting it was heading to a draw. But on the 55th move, Ding committed a fatal blunder – moving his rook into a position to be taken. Immediately recognising his mistake, he slumped on the table. "Ding seemed to have a risk-free chance to push for a win, but instead liquidated into a pawn-down endgame," Chess.com wrote in its post-game summary. "It should have been drawn, but Ding blundered as the pressure grew." He resigned three moves later. Dommaraju promptly burst into tears as the room erupted with cheers. "I probably got so emotional because I did not really expect to win that position," he said. At age 18, he is only the second Indian player to become world chess champion, after five-time world chess champion Viswanathan Anand who last won in 2012. "It's a proud moment for chess, a proud moment for India… and for me, a very personal moment of pride." Indian Prime Minister Narendra Narendra Modi was also among the first public figures showering praise. "Historic and exemplary!" he wrote on X. "Congratulations to Gukesh D on his remarkable accomplishment. This is the result of his unparalleled talent, hard work and unwavering determination." The FIDE World Chess Championship carries a $2.5m (£1.96m) prize fund.""",
    """Text 10 : The AI wars: How Amazon, Microsoft, Google and Meta are investing. Big technology giants are going all in on artificial intelligence, investing in their own internal labs producing AI assistants, or making major investments into external companies such as OpenAI. In an effort to achieve artificial general intelligence, Google procured Deep Mind, which developed the Gemini assistant, and Meta has its own Meta AI lab, which developed the Llama model. Microsoft has invested into OpenAI, best known for ChatGPT, and Amazon has invested US$8 billion into its competitor Anthropic, which has the chatbot Claude. But Amazon has also been quietly building its own AI models – it revealed six new models at its annual re:Invent conference in Las Vegas last week, include a video and image generator to compete with OpenAI’s Dall-E and Sora models. “We will continue to invest in this, because it’s such an exciting time but it’s also such early days in the technology cycle.” Amazon Web Services head of AI and data Swami Sivasubramanian told Markets with Madison in Las Vegas. “Nobody knows which model architectures are going to be durable in the long run. “Nobody knows which assistants are going to resonate or not resonate.” Learn about the AI wars between big technology giants, including why Amazon is investing across all layers of the AI stack, in today’s episode of Markets with Madison above. Get investment insights from executives and experts on Markets with Madison every Monday and Friday here on the NZ Herald, YouTube and wherever you get your podcasts. Sponsored by CMC Markets. Madison travelled to Las Vegas courtesy of Amazon Web Services. Disclaimer: The information provided in this programme is of a general nature, and is not intended to be personalised financial advice. We encourage you to seek appropriate advice from a qualified professional to suit your individual circumstances. Madison Reidy is host and executive producer of the NZ Herald’s investment show Markets with Madison. She joined the Herald in 2022 after working in investment, and has covered business and economics for television and radio broadcasters.""",
    """Text 11 : Canterbury farmer Chris Allen’s sudden death: Grieving family remembers loving husband and father. Mid-Canterbury farmer Chris Allen is remembered by many as a nationally acclaimed agriculturalist and a voice for nature – but his wife Anne-Marie simply remembers him as a “loving husband and a dedicated father”. Allen was 62 when he died on Monday after he was electrocuted while fixing a garage door. Close friend, Environment Canterbury’s Ian Mackenzie said there was “no predicting” what happened. Online tributes have described his extensive contributions to the agriculture industry, and his “great legacy”. But to Anne-Marie and his family, Allen was a “family man and a people person” above all else. He was a “loving husband and a dedicated father” to their two children, Natalie and Jonathan. “He was just a joker ... people wouldn’t know whether to believe what he was saying sometimes but his goal was always to make people smile,” Anne-Marie told the Herald. The pair met in Christchurch more than 30 years ago. Both from farming families originally, it wasn’t long before they had a farm of their own, Anne-Marie revealed. “After we met, he was quite determined we needed some land!” she said. The family spent summers camping by the river or holidaying in Mt Maunganui, all the while Allen spent his spare time giving back to the community, Anne-Marie said. “Chris was always heavily involved in local schools and committees ... because of his various roles including Federated Farmers and Air New Zealand, he had a lot of contacts,” she said. An engineer originally, Allen recently attended the 70-year anniversary of the engineering base at Air New Zealand in Christchurch. Anne-Marie described how Allen grew up with a passion for animals, farming and rivers. When the 2021 Mid Canterbury floods hit, they both set out to help the community – despite the fact their own farm was inundated with water. “The community needed help ... and he had the contacts to get the job done,” she said. Anne-Marie explained how no matter the situation, Allen “always said yes”. “He was solution-based, he never did it with confrontation – it was always about working with people to get the best outcome he could,” she said. Overall, Anne-Marie said “he loved his family and he loved his friends”. “He was the youngest of four and very close to his two remaining siblings that are alive,” she said. Mackenzie told the Herald he will remember Allen for his “wicked sense of humour”. He said the Ashburton community will be “absolutely gutted”. “He was just an outstanding fellow,” he said. “He was articulate and passionate about agriculture and the environment but above all else, he was just a lovely man.” Mackenzie recalled how they used to travel to Wellington together often for work trips. “Allen knew every single flight attendant by name, that’s just the kind of lovely man he was ... so if you ended up sitting next to Chris, I guarantee he would’ve chatted to you,” he said. Allen was a respected farmer in the Mid Canterbury region for years – and a long-time, vocal member of Federated Farmers. His family told the Herald they are “devastated and can’t really put it into words”. “We loved him for who he was and we are incredibly proud,” Anne-Marie said.""",
    """Text 12 : Brazil’s Lula to undergo additional medical procedure after brain bleed emergency surgery. Brazilian President Luiz Inacio Lula da Silva is expected to undergo an additional medical procedure on Thursday as part of his treatment following the emergency surgery he had Tuesday to drain a bleed on his brain. Medics from Sao Paulo’s Sírio-Libanês Hospital will perform an endovascular procedure on Thursday morning as part of his “therapy program,” according to a note released by the hospital Wednesday afternoon obtained by CNN Brasil. The president’s doctor, Roberto Kalil, told Reuters the procedure was considered a “complement” to Tuesday’s surgery. “This involves a catheterization that targets the meninges artery. Why? Because when a hematoma is drained, there is a slight chance that the small arteries in the meninges could cause a minor bleed in the future,” Kalil said. In previous comments soon after the surgery, Kalil had said it was “important to note that (Lula) did not sustain any brain injury. The hematoma, the bleeding, was not inside the brain; it was outside the brain and has been completely drained.” Lula – who remains in the hospital’s intensive care unit – spent the “day well, without complications, underwent physiotherapy, walked and received visits from family,” the note released by the hospital says, adding that “the president remains under the accompaniment of the medical team.” Related article Brazilian President Lula stable in intensive care after emergency brain surgery The 79-year-old Lula was operated on Monday due to a brain bleed linked to a fall in October. The emergency surgery added to health concerns about the aging president, a standard-bearer of the Latin American left who is halfway through his third non-consecutive term. Lula has curtailed travel in recent months while doctors monitored his recovery from trauma to the back of his head, sustained when he fell at his home in late October, requiring stitches. Prior to the emergency surgery, Lula had complained of a worsening headache during a congressional event Monday evening in the capital, Brasilia, and was taken to a local hospital for examination, presidential spokesperson Paulo Pimenta said in a radio interview. An MRI scan detected an intracranial hemorrhage, and Lula was flown to Sao Paulo soon after for the surgery,  according to a medical note released by the government.""",
    """Text 13 : Iwi leader Bayden Barber sick of freedom campers at Waimarama Beach ‘doing their business wherever’. Although there are no designated freedom camping areas at Waimarama Beach one resident says that doesn’t stop some campers. Photo / NZME There are no freedom camping spots at Waimarama Beach — but an iwi leader says that doesn’t stop campers “doing their business wherever”. Ngāti Kahungunu chairman and resident Bayden Barber said such campers were a big problem in summer. “They leave lots of mess where they camp. There are no toilets, so they are just doing their business wherever,” he said. “Council signage is not adequate and needs to be larger. There is no policing of these areas, which is why locals like me have to do it. “I often have to remind them that there is no camping in these areas (Tiakitai Rd where the rocks are).” He said the rocks had blocked off the worst place for freedom camping and there was a camping ground in Waimārama. A Hastings District Council spokesperson said there were no Hastings bylaws prohibiting freedom camping. “We don’t have general signs saying, ‘no freedom camping’ in areas that are not designated for self-contained vehicles. “The only rules in the parking space that would apply would be abandoned vehicles, which means council staff will check on a vehicle parked on a street without being moved for a minimum of seven days.” Hastings has designated sites for self-contained vehicles at Evers-Swindell Reserve, Clifton Rd Reserve, Haumoana Domain, Pākōwhai Country Park, Puketapu Domain, Maraetotara Falls carpark, and Dartmoor Bridge Reserve. In these freedom camping areas, signs advise what people are allowed to do. The spokesperson said in these areas people were allowed to stay for a maximum of two nights a month per site. “All dogs must be on leads, no fires can be lit and their current certified self-containment motor caravan/caravan warrant must be on display.” Within Hastings City, visitors can stay in specified public car parks for a maximum of two nights. Parking is free overnight (6pm to 8am), however, normal charges apply during the day (8am to 6pm). “Security contractors patrol Evers-Swindell Reserve, Haumoana Domain and Clifton Reserve nightly and report any issues to council. If visitors are not compliant with the rules a trespass notice can be issued.” The most common issue they dealt with was littering. “From time to time we also have to deal with people lighting fires and defecating in open spaces. “When this behaviour is reported to council we respond appropriately, potentially issuing a trespass notice for major or persistent issues.” They said in general the majority of freedom campers followed the rules. Wairoa also has freedom camping spots which thousands of people flock to over the summer. Wairoa District Council chief executive Kitea Tipuna said over the past two summer seasons, there had been very few freedom camping infringements issued. “We have staff monitoring the camping situation, and sometimes we may need to have an educational chat, and people may have to move on, but generally, we are impressed with the visitors who flock to our district,” Tipuna said. Respect was key. “As long as campers are compliant, self-aware of others and respectful of our whenua and infrastructure, then we welcome them to enjoy a holiday in our district.” Hot spots over the summer break are at Mahia. The council said this year, the Blue Bay site has been slightly restricted in size because of coastal erosion and in a bid to give campers more space. Freedom camping is national legislation that allows people to camp on public land that isn’t a recognised camping ground or holiday park. Freedom camping is permitted in any local authority area within the district unless it is restricted or prohibited.""",
    """Text 14 : Elton John believes legalizing marijuana ‘is one of the greatest mistakes of all time’. Elton John is not holding his tongue when it comes to legalized marijuana. In an interview with Time magazine as part of their Icon of the Year coverage, the legendary singer shared his thoughts about the legalization of cannabis in parts of the US and Canada. “I maintain that it’s addictive,” he said. “It leads to other drugs. And when you’re stoned—and I’ve been stoned—you don’t think normally.” “Legalizing marijuana in America and Canada is one of the greatest mistakes of all time.” John comes at it from the perspective of being sober for many years. “He is Eminem’s sponsor. He orchestrated English pop star Robbie Williams’ first stint in rehab,” the publication pointed out. One artist he regrets he couldn’t help was the late George Michael, who struggled with substance abuse before dying with heart and liver disease in 2016. He was 53. “It’s tough to tell someone that they’re being an a**hole, and it’s tough to hear,” John said of attempting to assist Michael with sobriety. “Eventually I made the choice to admit that I’m being an a**hole.” John’s longtime writing partner, Bernie Taupin, said that when John was hooked on drugs, “I was terrified for him. It was absolutely horrible.” “A lot of the work that we did in the times when he was at his worst wasn’t the best of both of us,” Taupin said. “I wasn’t able to creatively invest any time in writing material that related to him until he actually found himself, and then it was easier for me to reflect upon it.”""",
    """Text 15 : Killing of UnitedHealthcare CEO prompts flurry of stories on social media over denied insurance claims. The early morning killing of a top health insurance executive in midtown Manhattan Wednesday has unleashed a flurry of rage and frustration from social media users over denials of their medical claims, a public display of Americans’ pent-up anger at the nation’s complex health insurance industry. In one stark example, a Facebook post by UnitedHealth Group expressing sadness about UnitedHealthcare CEO Brian Thompson’s death received 62,000 reactions – 57,000 of them laughing emojis. UnitedHealth Group is the parent company of UnitedHealthcare, the division that Thompson ran. New York police have yet to determine the shooter’s identity and motive, but in an interview with NBC, Thompson’s widow, Paulette, told the network that “there had been some threats” against her husband, potentially related to “a lack of coverage… I don’t know the details.” A source told CNN that there were concerning threats against UnitedHealth Group, although Thompson’s name was not mentioned. Related article UnitedHealthcare CEO’s killing shows why companies spend millions to protect their top executives A shell casing recovered from two of the bullets at the scene of the shooting had the words “depose” and “delay” written on them, according to law enforcement sources. Police are exploring whether the words indicate a motive since they are similar to the title of a book that’s critical of the broader insurance industry – “Delay Deny Defend.” The majority of insured US adults had at least one issue, including denial of claims, with their health insurance in the span of a year, according to a survey released in June 2023 by KFF, a nonprofit health policy research group. “Limitations on access to care due to claims denials have absolutely been a source of frustration for a long time,” said Kaye Pestaina, KFF’s director of Program on Patient and Consumer Protections. A major industry group for health insurers criticized some of the social media reactions. “The people in our industry are mission-driven professionals working to make coverage and care as affordable as possible and to help people navigate the complex medical system,” Mike Tuffin, CEO of the trade association, AHIP, said in a statement. “We condemn any suggestion that threats against our colleagues – or anyone else in our country – are ever acceptable.” UnitedHealthcare said in a statement to CNN that its main priorities are to support Thompson’s family, ensure the safety of its employees and work with law enforcement to bring the gunman to justice. “We, at UnitedHealth Group, will continue to be there for those who depend upon us for their health care,” the statement said. Other insurers did not return requests for comment. Almost immediately after news broke that Thompson had been killed, social media users began posting about their frustrations with UnitedHealthcare and other insurance companies. UnitedHealthcare “denied my surgery two days before it was scheduled. I was in the hospital finance office in tears (when I was supposed to be at the hospital doing pre-op stuff),” one user wrote in an X post that received more than 70,000 likes. “My mother was flying out to see me. My surgeon spent a day and a half pleading my case to United when she probably should have been taking care of her other patients,” she added, before saying the surgery ended up going ahead but calling the process “torture.” “My breast cancer surgery was denied” by a different insurance company, another X user posted. “Breast cancer. She asked me ‘well, is it an emergency?’ I don’t know- it’s (f***ing) cancer. What do you think? I had to appeal and luckily it went through. Evil to do that to people,” she said. Their stories could not be independently verified by CNN. Related article How do you feel about your health insurer? Share your story with CNN TikTokker and anesthesiologist Brian Schmutzler said in a video that the shooting was “obviously tragic for him, tragic for his family, but this brought up some bigger issues.” “From my perspective, we have a bigger issue with the insurance companies in general, who, essentially, it’s their job to make money, not to actually pay for health care,” he said. In a blog post Thursday entitled “Why ‘we’ want insurance executives dead,” journalist Taylor Lorenz, who covers social media, analyzed the online responses: “No, that does not mean people should murder them. But if you’ve watched a loved one suffer and die from insurance denial, it’s normal to wish the people responsible would suffer the same fate.” Lorenz’s post sparked a wave of backlash online. It’s not uncommon for Americans and their doctors to go through multiple hoops to get approval for the care doctors say their patients need, or to combat denials from carriers – even though 81% of insured adults in KFF’s 2023 survey rated their health insurance as “excellent” or “good.” “This is care that people’s doctors recommend for them, and some of this care can be frightening,” Sara Collins, senior scholar at The Commonwealth Fund, a health policy foundation, told CNN, giving a cancer diagnosis as an example. “To get any kind of denial or delay while your insurer figures out whether or not they’re going to cover it is frightening for people. To have a decision all of a sudden be being made on the basis of financing is terribly upsetting for families.” Nearly one in five insured adults experienced claim denials during a 12-month period, according to KFF’s 2023 survey. Those with job-based insurance or Affordable Care Act policies ran into this problem about twice as often as those covered by Medicare or Medicaid, whose denial rates were around one in 10. Related video Security expert on connection between shell casing discovery and health care industry Restricting access to health care through tools like claim denials and prior authorization, which requires that insurers approve the care in advance, are among the ways that health insurers try to weed out care that’s not medically necessary or not backed by scientific evidence – but it can also increase their profit margins. The practices, which increasingly rely on technology, including artificial intelligence, can infuriate patients and providers alike. A class action lawsuit filed last year in US District Court in Minnesota argued that UnitedHealthcare uses AI “in place of real medical professionals to wrongfully deny elderly patients care,” according to the complaint. More than 90% of the denials are reversed through an internal appeal or proceedings before federal administrative law judges, the suit alleges. UnitedHealthcare asked the court to dismiss the lawsuit, claiming the plaintiffs must first exhaust the administrative appeal process set by the Medicare Act, among other reasons. The denial of claims can lead to major problems for patients, both in terms of care and their finances. Last year, UnitedHealthcare settled a case brought by a severely ill college student in Pennsylvania who claimed the company denied coverage for drugs determined necessary by his doctors, leaving him with a medical bill of more than $800,000. The lawsuit, which was chronicled by ProPublica, uncovered the lengths to which the insurer goes to reject claims, including burying medical reports and relying on rubberstamped recommendations from doctors paid by the company. Roughly a quarter of consumers whose claims insurance companies denied experienced significant delays in getting medical care or treatment, and about the same share were unable to receive care, the KFF survey found. About a quarter of those said their health declined. Having to postpone care because of coverage denials can also affect patients’ mental health. Some 80% of adults said they or a family member were worried or anxious about delays in care, according to a 2023 Commonwealth Fund survey of adults ages 19 to 64. Relatively few people try to fight when their insurer turns down their claim. Only 43% of adults in the Commonwealth Fund survey said they or their doctor challenged an insurer’s denial of claim. Some 45% of respondents said they weren’t sure they had the right to appeal the denial, while 40% said they weren’t sure whom to contact. Nearly a quarter said they didn’t have the time. But challenging the denial can produce results. Half of those who appealed ultimately had their care approved, the Commonwealth Fund survey found. Inappropriate denials of services and payments by Medicare Advantage insurers – including UnitedHealthcare, the largest player in the swiftly growing market – have come under fire in recent years, particularly from the Department of Health and Human Services, which regulates the program, and from some lawmakers. The insurers, which are paid by the federal government to provide Medicare services to enrollees, have at times delayed or denied beneficiaries’ access to medical care – even though the requests met Medicare coverage rules, according to a 2022 report from HHS’ inspector general’s office. Annual federal audits have highlighted “widespread and persistent problems related to inappropriate denials of services and payment,” the office said. A central concern is the potential incentive Medicare Advantage plans have to deny access to services and payments to providers in an attempt to increase profits, the office said. Insurers are given a set amount of money per patient regardless of the amount of care received. UnitedHealthcare, in particular, has come under public scrutiny as it dramatically increased care denials for its Medicare Advantage enrollees. The insurer more than doubled the rate of denials for care following hospital stays between 2020 and 2022 as it implemented machine-assisted technology to automate the process, according to a Senate Permanent Subcommittee on Investigation’s report released in October. That far surpassed its competitors, including Humana, whose care denials grew 54% during the same time period. “Despite alarm and criticism in recent years about abuses and excesses, insurers have continued to deny care to vulnerable seniors—simply to make more money,” Connecticut Sen. Richard Blumenthal, who chairs the subcommittee, said in a statement in October."""
]

prompts = [
    """Prompt 1 : 
#INSTRUCTION :
Classifies whether a text is appropriate for an audience under 15 years old, considering the defined #SCALE.
Strictly follow the #EXPECTECD OUTPUT estructure. Do not give any other explanation.

#SCALE :
 - 90% - 100%: Highly appropriate; text is completely suitable for audiences under 15.
 - 70% - 89%: Mostly appropriate; minor content may require parental guidance.
 - 50% - 69%: Partially appropriate; supervision is recommended.
 - 30% - 49%: Inappropriate; no recommended for audiences under 15 years old.
 - 0% - 29%: Highly inappropriate; unsuitable for audiences under 15 years old.
 
#EXPECTED OUTPUT:
**Appropriateness**: [Calculated Percentage] 
**Suggestion**: [ - Highly appropriate; text is completely suitable for audiences under 15.
 - Mostly appropriate; minor content may require parental guidance.
 - Partially appropriate; supervision is recommended.
 - Inappropriate; no recommended for audiences under 15 years old.
 - Highly inappropriate; unsuitable for audiences under 15 years old.]
**Analysis**: [provide a concise evaluation explaining the reasons for the score in one paragraph]
    """,
    
    """Prompt 2: 
#INTRUCTION
Classifies whether a text is appropriate for an audience under 15 years old.
Follow these chain of thoughts to reach the classification:

- Divide the text into the three main dimensions: Explicit Content (Violence and Sexual content), Text Complexity (Language and Mature Themes), and Sensitivity (Cultural and Emotional)
- Assess the potential impact of such content on an audience under 15 years old.
- Decide if the text meets the standards for content appropriate for this audience.

#EXPECTED OUTPUT:
**Appropriateness**: [Calculated Percentage] 
**Suggestion**: [ - Highly appropriate; text is completely suitable for audiences under 15.
 - Mostly appropriate; minor content may require parental guidance.
 - Partially appropriate; supervision is recommended.
 - Inappropriate; no recommended for audiences under 15 years old.
 - Highly inappropriate; unsuitable for audiences under 15 years old.]
**Analysis**: [provide a concise evaluation explaining the reasons for the score in one paragraph]
    """,
    
    """Prompt 3:
#PRIMARY OBJECTIVE
You are a specialised content evaluator designed to analyse and classify textual content for appropriateness for audiences under 15 years old. Your goal is to assess the text against pre-defined criteria, calculate a scoring percentage of appropriateness, and provide a structured output.

#FEATURES 
- Professional Role: Content Appropriateness Evaluator.  
- Core Capabilities:
  1. Detect explicit violent, sexual, mature themes, emotionally distressing, or culturally sensitive content.
  2. Assess the potential impact of the text on an audience under 15 years old.
  3. Rate the text with an "appropriateness" percentage based on defined #SCALE.
- Prioritize accuracy and consistency over creativity. 

#INSTRUCTIONS:
1. Receive a text provided by the user.
2. Analyze the content using the **SCORING SYSTEM AND CRITERIA**  defined below.
3. Calculate an "appropriateness" percentage based on the predefined #SCALE
4. Do not suggest edits or adjustments to the provided text. Your only role is to evaluate.

#SCALE :
 - 90% - 100%: Highly appropriate; text is completely suitable for audiences under 15.
 - 70% - 89%: Mostly appropriate; minor content may require parental guidance.
 - 50% - 69%: Partially appropriate; supervision is recommended.
 - 30% - 49%: Inappropriate; no recommended for audiences under 15 years old.
 - 0% - 29%: Highly inappropriate; unsuitable for audiences under 15 years old.

#EVALUATION CRITERIA
The scale is based on three core dimensions, each contributing equally to the final appropriateness score (33.3% each). These dimensions are Explicit Content, Textual Complexity, and Sensitivity.
1. Explicit Content (33.3%)
This dimension evaluates the degree of violence and sexual content in the text. Explicit or suggestive material can negatively impact the emotional well-being of young readers.
    1.1. Violence: The intensity, frequency, and graphic nature of violent descriptions are considered.
    Example: 
        - Severe violence with graphic details: -20%
        - Mild references to conflict without explicit descriptions: -5%
    1.2. Sexual Content: The presence of descriptions or references to sexual acts, nudity, or innuendo.
    Example:
        - Explicit descriptions or suggestive material: -20%
        - Subtle or indirect mentions: -5%
2. Textual Complexity (33.3%)
This criterion focuses on the accessibility of language and the maturity of the themes presented, ensuring alignment with the cognitive abilities of readers under 15.
    2.1. Language: Evaluates the use of advanced vocabulary, jargon, or language unfamiliar to younger audiences.
    Example
        - Extensive use of complex terms and difficult phrasing: -10%
        - Occasional challenging words: -5%
    2.2. Mature Themes: Considers topics such as substance abuse, severe illnesses, and social issues that may be inappropriate or confusing for young readers.
    Example
        - Detailed exploration of mature themes: -20%
        - Surface-level mentions without depth: -5%
3. Sensitivity (Cultural and Emotional) (33.3%)
This dimension assesses whether the content respects cultural diversity and protects readers from unnecessary emotional distress.
    3.1. Cultural Sensitivity: Representation of diverse groups without stereotypes or bias
    Example
        - Stereotypical, offensive, or exclusionary depictions: -20%
        - Neutral and inclusive representation: +5%
    3.2. Emotional Impact: Potential for the content to cause distress, fear, or confusion.
        - Traumatic or emotionally intense scenarios: -20%
        - Mild emotional challenges with positive resolutions: +5%

#EXPECTED OUTPUT
Always respond in the following structure:

    **Appropriateness**: [Calculated Percentage] 
    **Suggestion**: [ - Highly appropriate; text is completely suitable for audiences under 15.
     - Mostly appropriate; minor content may require parental guidance.
     - Partially appropriate; supervision is recommended.
     - Inappropriate; no recommended for audiences under 15 years old.
     - Highly inappropriate; unsuitable for audiences under 15 years old.]
    **Analysis**: [provide a concise evaluation explaining the reasons for the score in one paragraph]


#COMMUNICATION STYLE
- Tone: Professional, concise, and neutral.  
- Clarity: Technical yet accessible.


#NEGATIVE PROMPTS 
- Do not provide suggestions or explanations outside the analysis scope.  
- Do not deviate from the scoring system or the provided methodology.
  
    """
]

#Creating structure
data = {
    "texts" : texts,
    "prompts" : prompts
}

#Saving in a JSON file

with open("texts_and_prompts.json", "w") as file:
    json.dump(data, file, indent=4)

print("JSON file 'texts_and_prompts.json' created succesfully.")


from openai import OpenAI
import os

client = OpenAI(
    api_key=os.environ.get("OPENAI_API_KEY"))

#API confing
with open("texts_and_prompts.json", "r") as file:
    data = json.load(file)
    
texts = data["texts"]
prompts = data["prompts"]

#Models to tests
models = ["gpt-4o", "gpt-4o-mini", "o1-mini", "gpt-4-turbo"]

#Initialazing list to store results
results = []

#Combinations loop
for model in models:
    for prompt in prompts:
        for text in texts:
            try:
                #Model input
                message = prompt + text

                #OpenAI API calling
                if model == "o1-mini":
                    response = client.chat.completions.create(
                        model=model,
                        messages=[{"role": "user", "content": message}]
                        )
                else:
                    response = client.chat.completions.create(
                        model=model,
                        messages=[{"role": "user", "content": message}],
                        temperature=0.5
                        )

                # storing results
                results.append({
                    "model": model,
                    "prompt": prompt,
                    "text": text,
                    "output": response.choices[0].message.content
                })

            except Exception as e:
                #Ignoring errors
                results.append({
                    "model": model,
                    "prompt": prompt,
                    "text": text,
                    "error": str(e)
                })

print("Iteration completed successfully")

# Saving results in JSON file
with open("Testing_results.json", "w") as file:
    json.dump(results, file, indent=4)

#Dataframing results
df = pd.DataFrame(results)
df.to_csv("Testing_results.csv", index=False)

print("Testing completed. Results saving in 'Testing_results.json' and 'Testing_results.csv'.")

df

df.to_excel("Testing_results.xlsx", index=False)
