==> Testing Phase/04 - Analysis Updated.ipynb <==
import pandas as pd

#Reading data
Testing_results = pd.read_csv('Testing_results.csv')
Expected_output = pd.read_csv('Expected_Outout.csv')

# Extracting Prompt and text IDs
Testing_results['actual_suggestion'] = Testing_results['output'].str.extract(r'\*\*Suggestion\*\*: ?-? ([^;]+)').astype(str)
Testing_results['text_id'] = Testing_results['text'].str.extract(r'(Text\s*\d+)')
Testing_results['prompt_id'] = Testing_results['prompt'].str.extract(r'(Prompt\s*\d+)')
Testing_results

#Checking uniques data
print(Testing_results['actual_suggestion'].unique())
print(Expected_output['top'].unique())
print(Testing_results['prompt_id'].unique())
print(Testing_results['text_id'].unique())
print((Expected_output)['text'].unique())

Expected_output['text'] = Expected_output['text'].str.strip()
df = pd.merge(Testing_results, Expected_output[['text','top']], left_on='text_id', right_on='text', how='left')
df

#Checking data
df.info()

#Applying Label encoding to "actual_suggestion" and "top" ("expected_suggestion")
label_mapping = {
    'Highly appropriate':0 ,
    'Mostly appropriate':1,
    'Partially appropriate':2,
    'Inappropriate':3,
    'Highly inappropriate':4
}

df['y_true'] = (df['top'].str.strip()).map(label_mapping)
df['y_pred'] = (df['actual_suggestion'].str.strip()).map(label_mapping)

df.head(5)


#Final data to analise
final_data = df[["text_id", "prompt_id", "model", "y_true", "y_pred"]]
final_data.to_csv('final_data.csv')
final_data.to_excel('final_data.xlsx')

final_data.head()


final_data.info()

#Importing evaluation metrics library
from sklearn.metrics import (
    accuracy_score,     
    precision_score,    
    recall_score,        
    f1_score,            
    classification_report 
)

#Grouping by Model and Prompt
grouped = final_data.groupby(['model', 'prompt_id'])

#Function to compute metrics by group
def computing_metrics(group):
        
    return pd.Series({
        'Accuracy': accuracy_score(group['y_true'], group['y_pred']),
        'F1-Score': f1_score(group['y_true'], group['y_pred'], average='weighted'),       
        'Precision': precision_score(group['y_true'], group['y_pred'], average='weighted'),
        'Recall': recall_score(group['y_true'], group['y_pred'], average='weighted'),
        'Total_Samples': len(group)
    })

# Aplying Metrics to every group combination
Metric_results = grouped.apply(computing_metrics).reset_index()
Metric_results
